+++ history.rst
@@ -9,7 +9,11 @@ Version 0.0.3
 
 2015-09-02
 
-- API: Added *get_info()*-Function
+- JSON-RPC-API:
+
+  - Added *get_info()*-Function
+
+  - Create-Function finished and tested
 
 
 =============